+++ packages/expo-router/plugin/src/withRouter.ts
@@ -25,6 +25,7 @@ const withExpoHeadIos: ConfigPlugin = (config) => {
 
 const withGammaScreens: ConfigPlugin = (config) => {
   return withPodfile(config, (config) => {
+    // TODO: Remove after fully migrating to react-native-screens v5, which removed this flag.
     if (!config.modResults.contents.includes('RNS_GAMMA_ENABLED')) {
       config.modResults.contents = `# Set by expo-router. This enables Fabric-only features from react-native-screens\nENV['RNS_GAMMA_ENABLED'] ||= '1'\n${config.modResults.contents}`;
     }